split `tester` module into multiple
